requirements.txt
python-dotenv
google-genepip 
streamlit
